Version 4.0.6
1 - Has been remove the "elevate.exe" from te installer. Now in order to
add or remove the widget from the system startup, it uses the a windows
registry value. It is the way to do that and it does not require windows
UAC. When install the widget, creates the registry value, when uninstall
the widget, the installer will remove it if it exists. From the settings
view, if is the option cheked, the registry value will be created, if
you uncheck it, the value will be removed.
2 - Has been removed the shortcut from the startup folder.
3 - Has been removed the shortcut from the installation folder.